Lazy objects: Update class constants earlier
If a lazy object is created for a class whose constants can not be updated, then
we have created an instance of a non-instantiable class. This breaks the
expectations of `clone`.

Here I ensure that a class has its constants updated before creating a lazy
instance of it.

Fixes oss-fuzz #71407