Fix bug #45282: SoapClient has namespace issues when WSDL is distributed
We're ignoring the namespaces that the messages, ports, bindings and
services are defined in when registering them. Consequently, we're also
ignoring the namespace component of the QName when looking up
references.

This patch solves this issue by introducing a new qname_hash_table type
along with two methods to add and query the table. Internally, it's a
nested HashTable where the outer layer is the namespace uri and the
inner layer is the local name.

It's worth pointing out that at least one comment in the original issue
points to a WSDL that actually has a validation issue because it has two
messages with the same local name in the same namespace, so in that case
PHP's SOAP client is actually correct. The other links in the comments
no longer exist, but I was able to reconstruct the issue based on
reading the report and by cross-checking the expected behaviour with the
WSDL spec.