Bind trait props/consts before parent class
This more accurately matches the "copy & paste" semantics described in the
documentation. The same could be done for methods. However, abstract trait
methods diverge from this behavior, given that a parent method can satisfy trait
methods used in the child. In that case, the method is not copied, but the check
must be performed after the parent has been bound. For now, just bother with
properties and constants.

Fixes GH-15753