Remove Readline library integration
This removes the GNU Readline library integration and relies on the
libedit library if present. Issue is that there are several ext/readline
libraries and GNU Readline license (GNU GPL 3) is not compatible with
the PHP license. There was a similar attempt to fix this (GH-3823) and
now fix continues here.

Changes:
- `--with-libedit` configure option removed (only `--with-readline`
  stays)
- readline extension tests fixed: those required by only readline
  library removed, added missing skip reasons, some functions are always
  available, etc.
- New PHP_SETUP_EDIT Autoconf macro that finds libedit

This also fixes the phpdbg readline integration on *nix systems (using
libedit library) for history (up/down keys and similar nice features) to
not produce compile errors:

    ./configure --enable-phpdbg-readline
    make

    undefined reference to 'readline'
    undefined reference to 'add_history'