[skip ci] Drop `ZIP_STATIC` hack
Old versions of libzip didn't export the `ZIP_STATIC` macro, even if
they had been built as static libraries.  This hack in config.w32 is
supposed to counteract that, since the macro is used in zip.h.
However, PHP has no business in defining macros belonging to external
libraries; instead the winlibs builds should have been fixed to set
this macro in zip.h.  Furthermore, as of libzip there is zipconf.h,
where this macro is defined as appropriate, if libzip was build with
the proper configuration options.  Unfortunately, that has not been
done so far, but assuming that new properly configured builds of libzip
will be rolled out, we need to drop this hack to avoid conflicting
macro definitions, since we define the macro to 1, but libzip only
defines it (without further hackery).