Fix regression on systems built without JIT
regressing commit: 654b787ee16ae6ebf9ee94507631da9251ff077b

This was called if JIT was enabled or not. If not enabled, it'll result
in an undeclared function warning and maybe a bad time in the linker.
Gate the meat of this PHP-side function on if JIT is enabled (but keep
it existing so PHP userland code works with or without JIT, OFC).