Autotools: Use runstatedir for var/run directory
This removes duplicate substituted variables in FPM SAPI. The
php_fpm_sysconfdir is the same as EXPANDED_SYSCONFDIR and
php_fpm_localstatedir is the same as EXPANDED_LOCALSTATEDIR, which are
already done in the configure.ac.

Instead of hardcoding the var/run directory, the runstatedir can be
used (configured also via the `--runstatedir` configure option). By
default, the runstatedir is `${localstatedir}/run`.