Fix failing soap tests on Windows
These failures are caused by the fix for GHSA-p99j-rfp4-xqvq.  Since
the two bug*.phpt tests don't need the "wsdl" query string, we use a
`--CGI--` section instead.  The two server*.phpt tests are harder to
fix, since during evaluation of the `--SKIPIF--` section, the soap
extension can be loaded, but it may not during evaluation of the
`--FILE--` section.  So for now, we skip these tests on Windows
altogether.