Update Windows CI to use php-sdk-2.3.0
php-sdk-2.2.0 still fetches dependencies from the no longer up to date
<https://windows.php.net/downloads/php-sdk/deps/>, and as such won't be
tested with any security updates we provide for Windows.  Given that
PHP 8.1 is going to receive security updates for further 15 months, we
should should not ignore these dependency updates.

We also fix failing OpenSSL tests, which are no longer failing, at
least on Windows, because we have back-ported the fix for the Marvin
Attack[1].  So we fix the test cases accordingly.

[1] <https://github.com/php/php-src/security/advisories/GHSA-hh26-4ppw-5864>