Re-introduce caching for Windows CI
This had been done when running on AppVeyor, but had been dropped when
Windows CI was migrated to Github.  While not terribly useful regarding
the performance (maybe a minute can be saved), it still appears to be
reasonable to avoid unnecessary requests to downloads.php.net.

To have the branch available in the action, we make detection of the
target branch a separate step.