Fix nightly builds regarding libavif
Currently, we're running LINUX_X64 Asan builds on ubuntu-20.04, where
libavif-dev is not available, so apt-x64 fails.  This PR is a follow up
to PR #16049 to pass the respective parameter to apt-x64, and to fix
the conditional check in configure-x64.

It might be clearer to just have an `avif` parameter or to pass the
runner version to these actions, but this should do for now.