[skip ci] Support running of a selected nightly job
For debugging and development purposes it is occasionally useful to
one run one of the nightly jobs manually.  We support that by adding a
choice input where the desired job can be selected (`*` means run all
jobs).

Note that we generate the matrix even when it's not required, and that
we do not support constraining the matrix for now.