Fix regression on platforms without ZEND_CHECK_STACK_LIMIT set
There was a guard on one instance of this, but not the other. Unsure if
this is the best place for this; the functions are `#ifdef` gated in
`zend_execute.c`, but not the associated header; perhaps it should be
`#else` with a nop version of the function in that case? I added a
guard at this call site to be consistent with other places.