Install MySQL manually instead of using Chocolatey
While Chocolatey simplifies things, we cannot cache the installs, and
as such need to download and setup over and over again.  Particularly,
the MySQL installation fails sometimes for nightly builds, so we're
better off using a cache.

We also switch to MariaDB, because the downloads are much smaller so we
safe some storage space for other caches.