Reflection: show the type of object constants used as default properties
When a property default is based on a global constant, show the type of the
default. Previously, `format_default_value()` assumed that non-scalar and
non-array defaults were always going to be `IS_CONSTANT_AST` pointers, and when
the AST expression had been evaluated and produced an object, depending on when
the `ReflectionClass` or `ReflectionProperty` instance had been created, the
default was shown as one of `callable`, `__CLASS__`, or `...`.

Instead, if the default value is an object (`IS_OBJECT`), show the type of that
object.

Add test cases for each of the `callable`, `__CLASS__`, and `...` cases to
confirm that they all now properly show the type of the constant.

Closes gh-15902