Push NDEBUG handling down to zend_portability.h
As is, `NEDBUG` is set to match `PHP_DEBUG` (which matches `ZEND_DEBUG`)
in php.h right before including assert.h; however, assert.h is already
included at this point, so the `NDEBUG` handling is useless.  We push
it down to zend_portability.h, where it has effect.

The alternative would be to drop that code in php.h altogether, but
that would make some test assumptions invalid.   E.g. a build with
`CFLAGS=-DNDEBUG ./configure --enable-debug-assertions` would fail
internal_functions001.phpt, because the userland `PHP_DEBUG` is true,
but assertions would not be evaluated.