Move ARG_(WITH|ENABLE) to the toplevel
`buildconf` (and `phpize`) have special treatment for these "macros".
When configure.js is built, all config.w32 are grepped, these "macros"
are appended to configure.js, and all config.w32 contents are appended
with the "macros" commented out.  That means that for `configure` they
are in the toplevel anyway, so having them inside of `if` statements in
config.w32 is confusing.

Note that this matches autoconf behavior.