Use ZendMM for libgmp
As is, ext/gmp uses the default memory allocator of the underlying
libgmp; this allocator is fallible, but contrary to ZendMM, it
`abort()`s without giving us an opportunity to do a cleaner shutdown.
Furthermore, the libgmp allocator obviously doesn't heed memory_limit.
Thus we install wrappers of the ZendMM API as libgmp allocator.