Use ZendMM in ext-gmp
While libgmp uses true globals to track the allocator, and is therefore
not safe to use in a threaded environment, we can make use of a
thread-local variable that indicates whether the current thread is
currently performing a PHP request. We set this TLS global to true if it
is in RINIT, and set it to false in post-RSHUTDOWN. This ensures that
any (de)allocation during the request will go through ZendMM while
keeping it possible for this thread to be reused or for other threads
using libgmp.

Co-authored-by: "Christoph M. Becker" <cmbecker69@gmx.de>